split ProposalPolicy::Scope specs to their own file
